+++ Runtime/SerializationExtensions.cs
@@ -84,11 +84,11 @@ namespace Unity.RuntimeSceneSerialization
         internal static decimal AsDecimal(this UnsafePrimitiveView view)
         {
             return decimal.Parse(view.AsString(), System.Globalization.NumberStyles.Any, System.Globalization.CultureInfo.InvariantCulture);
-        }
+    }
 
         internal static float AsFloatSafe(this UnsafePrimitiveView view)
         {
             return float.Parse(view.AsString(), System.Globalization.NumberStyles.Any, System.Globalization.CultureInfo.InvariantCulture);
-        }
+}
     }
 }