+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.0-preview] - 2022-05-12
+### Changed
+- Properties and Serialization dependency version to 1.8.2-preview
+
+## [0.5.0-preview] - 2022-05-12
+### Changed
+- Properties and Serialization dependency version to 1.7.0-preview
+
+## [0.4.0-preview] - 2022-05-12
+### Changed
+- Minimum Editor version to 2020.1.0f1
+- Properties and Serialization dependency version to 1.6.0-preview
+
 ## [0.3.5-preview] - 2022-05-05
 ### Added
 - SceneSerialization.FromJsonOverride method for deserializing to existing containers